Jenkinsfile

+++ Jenkinsfile
@@ -21,21 +21,21 @@ pipeline {
             steps {
                 echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
                 script {
-                if (${params.CLEAN_BUILD}) {
-                    deleteDir()
-                }
+                    if (${params.CLEAN_BUILD}) {
+                        deleteDir()
+                    }
                 }
                 checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])				
                 sh 'autobuild/checkout-submodule.sh ${BC_LIB}'
                 sh 'autobuild/_brainCloudSetup_examples-unreal.command'
                 sh 'autobuild/makebuild.sh RelayTestApp MAC UE_5_Mac'
-                //sh 'autobuild/makebuild.sh RelayTestApp IOS'
-                //sh 'autobuild/makebuild.sh RelayTestApp ANDROID'
-/*                 sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh TappyChicken MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh ScriptTestApp MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh Groups MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac' */
+                sh 'autobuild/makebuild.sh RelayTestApp IOS UE_5_IOS'
+                //sh 'autobuild/makebuild.sh RelayTestApp ANDROID UE_5_Android'
+                //sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
+                //sh 'autobuild/makebuild.sh TappyChicken MAC UE_5_Mac'
+                //sh 'autobuild/makebuild.sh ScriptTestApp MAC UE_5_Mac'
+                //sh 'autobuild/makebuild.sh Groups MAC UE_5_Mac'
+                //sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac'
             }
             post {
                 success {
@@ -56,22 +56,24 @@ pipeline {
               }
              steps {
                  echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
-                 //if (${CLEAN_BUILD}) {
-                    deleteDir()
-                 //}
+                 script {
+                    if (${params.CLEAN_BUILD}) {
+                        deleteDir()
+                    }
+                 }
                  checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
                  bat 'autobuild\\_brainCloudSetup_examples-unreal.bat'
-                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win'
-                 //bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
-                 //b/at 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat Groups Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_4_Win'
+                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win64'
+                 //bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_5_Win64'
+                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_5_Win64'
+                 //b/at 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_5_Win64'
+                 //bat 'autobuild\\makebuild.bat Groups Win64 UE_5_Win64'
+                 //bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_5_Win64'
             }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win64', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
@@ -87,20 +89,22 @@ pipeline {
               }
              steps {
                  echo "---- braincloud Code Pull ue4-examples ${BC_LIB}"
-                 //if (${CLEAN_BUILD}) {
-                     deleteDir()
-                 //}
+                 script {
+                    if (${params.CLEAN_BUILD}) {
+                        deleteDir()
+                    }
+                 }
                  checkout([$class: 'GitSCM', branches: [[name: '*/ue4-examples']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
                  bat 'autobuild\\_brainCloudSetup_examples-unreal-4.bat'
-                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win'
-                // bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
-                 //bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
+                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win64'
+                // bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win64'
+                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win64'
+                 //bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win64'
              }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win64', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
